chore: improve Cloud E2E Test speed
